created the first version of the project with the following files:
    Bell_Curve.py
    Gamma_Distribution.py
    Lognormal_Distribution.py
    Main.py
    Math.py
    Normal_Distribution.py
    Test.py

which includes the distributions bell curve, gamma, lognormal, and normal distributions. The Math.py file includes the functions for the mean, standard deviation, and variance. The Main.py file includes the main code to run the project. The Test.py file includes the tests for the distributions.

Changes to be committed:
	new file:   Bell_Curve.py
	new file:   Gamma_Distribution.py
	new file:   Lognormal_Distribution.py
	new file:   Main.py
	new file:   Math.py
	new file:   Normal_Distribution.py
	new file:   Test.py
	new file:   __pycache__/Bell_Curve.cpython-311.pyc
	new file:   __pycache__/Gamma_Distribution.cpython-311.pyc
	new file:   __pycache__/Lognormal_Distribution.cpython-311.pyc
	new file:   __pycache__/Math.cpython-311.pyc
	new file:   __pycache__/Normal_Distribution.cpython-311.pyc